initialized git and now my progress will be tracked
This dictionary app is not your typical dictionary app. It will have a typical search for the definitions of words, but it will also have a reverse search with a built-in dictionary that I will make with as many coding terms and keywords as I can type, lol. But that is not all, it will also give a code snippet of the term whenever needed. This is an app that I can constantly update with new terms, and I believe it will make learning React an easier process.