added a dictionary component with form and search engine
Added all logic to handle the change of keyword from empty to whatever the user adds with useState, then added the logic to handle the onChange and store it with the useState, and then a search function to search for the keyWord, but for right now, I have set the search to an alert until I integrate the Api. I have also added a console log to let me know that the keyword is being searched.