finalized the api calls. Finished with all styling
I could not get the Pexels API to run. I was getting CORS errors, so I made a separate proxy server.js, which worked, but I am deploying a single page to Netlify and could not do that with the proxy server.js in a second folder. I could have used a free backend, but I did not because of classroom purposes.